requirements.txt
pandas
requests
whoisdomain
python-whois
dnspython
mysql-connector-python
python-dotenv